# apps/client-dashboard/public/openapi.yaml
openapi: 3.0.3
info:
  title: Markova API
  version: 1.0.0
  description: 'API-first voice and workflow agent platform (Domain 2).

    Sandbox keys (`mk_test_*`) and live keys (`mk_live_*`) share the same shapes;

    side effects differ (no real telephony spend in sandbox).

    '
servers:
- url: https://api.markova.tech
  description: Production API Gateway (Cloudflare Edge)
- url: https://markova-api-gateway.onrender.com
  description: Direct Render Gateway
- url: http://localhost:8000
  description: Local Development Gateway
...
security:
- bearerAuth: []
- apiKeyAuth: []
paths:
...
  /v1/agents/{id}/test-call:
    post:
      tags:
      - Agents
      summary: Sandbox-only test call (no billing)
      description: Requires a `mk_test_` key. Rejected for live keys.
      parameters:
      - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - to_number
              properties:
                to_number:
                  type: string
                  example: '+251911000000'
      responses:
        '201':
          description: Sandbox call placed
        '403':
          description: Live key not allowed
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: Tenant API key (`mk_test_*` or `mk_live_*`)
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...