# services/api-gateway/src/app.controller.ts
import { Controller, All, Req, Res, Get } from '@nestjs/common';
import { Request, Response } from 'express';
import * as path from 'path';
import * as fs from 'fs';
import { proxyTo } from './proxy.util';

@Controller()
export class AppController {
  private authServiceUrl = process.env.AUTH_SERVICE_URL || 'http://auth-service:5001';
  private tenantServiceUrl = process.env.TENANT_SERVICE_URL || 'http://tenant-service:5002';
  private agentBuilderUrl = process.env.AGENT_BUILDER_URL || 'http://agent-builder:5003';
  private toolEngineUrl = process.env.TOOL_ENGINE_URL || 'http://tool-engine:5004';
  private orchestratorUrl = process.env.ORCHESTRATOR_URL || 'http://orchestrator:6000';
  private connectorHubUrl = process.env.CONNECTOR_HUB_URL || 'http://connector-hub:5005';
  private knowledgeServiceUrl = process.env.KNOWLEDGE_SERVICE_URL || 'http://knowledge-service:5006';
  private capabilityRegistryUrl = process.env.CAPABILITY_REGISTRY_URL || 'http://capability-registry:5009';
  private plannerServiceUrl = process.env.PLANNER_SERVICE_URL || 'http://planner-service:5011';

  // ── Legacy /api/* ──────────────────────────────────────────────────────────

  @All('api/auth*')
  proxyAuth(@Req() req: Request, @Res() res: Response) {
    return proxyTo(this.authServiceUrl, req, res);
  }

  @All('api/clients*')
  proxyClients(@Req() req: Request, @Res() res: Response) {
    return proxyTo(this.authServiceUrl, req, res);
  }

  @All('api/users*')
  proxyUsersLegacy(@Req() req: Request, @Res() res: Response) {
    return proxyTo(this.authServiceUrl, req, res);
  }

  @All('api/invitations*')
  proxyInvitationsLegacy(@Req() req: Request, @Res() res: Response) {
    return proxyTo(this.authServiceUrl, req, res);
  }

  @All('api/roles*')
  proxyRolesLegacy(@Req() req: Request, @Res() res: Response) {
    return proxyTo(this.authServiceUrl, req, res);
  }

  @All('api/departments*')
  proxyDepartmentsLegacy(@Req() req: Request, @Res() res: Response) {
    return proxyTo(this.authServiceUrl, req, res);
  }

  @All('api/sessions*')
  proxySessionsLegacy(@Req() req: Request, @Res() res: Response) {
    return proxyTo(this.authServiceUrl, req, res);
  }

  @All('api/workspace*')
  proxyWorkspaceLegacy(@Req() req: Request, @Res() res: Response) {
    return proxyTo(this.authServiceUrl, req, res);
  }

  @All('api/tenant*')
  proxyTenant(@Req() req: Request, @Res() res: Response) {
    return proxyTo(this.tenantServiceUrl, req, res);
  }

  @All('api/contact')
  proxyContact(@Req() req: Request, @Res() res: Response) {
    return proxyTo(this.tenantServiceUrl, req, res);
  }

  @All('api/crm*')
  proxyCrm(@Req() req: Request, @Res() res: Response) {
    return proxyTo(this.tenantServiceUrl, req, res);
  }

  @All('api/builder*')
  proxyBuilder(@Req() req: Request, @Res() res: Response) {
    return proxyTo(this.agentBuilderUrl, req, res);
  }

  @All('api/tools*')
  proxyTools(@Req() req: Request, @Res() res: Response) {
    return proxyTo(this.toolEngineUrl, req, res);
  }

  @All('api/connectors*')
  proxyConnectors(@Req() req: Request, @Res() res: Response) {
    return proxyTo(this.connectorHubUrl, req, res, (url) =>
      url.replace(/^\/api\/connectors/, '/api/connector-hub'),
    );
  }

  @All('api/connector-hub*')
  proxyConnectorHub(@Req() req: Request, @Res() res: Response) {
    return proxyTo(this.connectorHubUrl, req, res);
  }

  @All('api/knowledge*')
  proxyKnowledge(@Req() req: Request, @Res() res: Response) {
    return proxyTo(this.knowledgeServiceUrl, req, res);
  }

  // ── Canonical /v1/* (Phase 2) ──────────────────────────────────────────────

  @All('v1/auth*')
  proxyAuthV1(@Req() req: Request, @Res() res: Response) {
    return proxyTo(this.authServiceUrl, req, res);
  }

  @All('v1/users*')
  proxyUsersV1(@Req() req: Request, @Res() res: Response) {
    return proxyTo(this.authServiceUrl, req, res);
  }

  @All('v1/invitations*')
  proxyInvitationsV1(@Req() req: Request, @Res() res: Response) {
    return proxyTo(this.authServiceUrl, req, res);
  }

  @All('v1/roles*')
  proxyRolesV1(@Req() req: Request, @Res() res: Response) {
    return proxyTo(this.authServiceUrl, req, res);
  }

  @All('v1/departments*')
  proxyDepartmentsV1(@Req() req: Request, @Res() res: Response) {
    return proxyTo(this.authServiceUrl, req, res);
  }

  @All('v1/sessions*')
  proxySessionsV1(@Req() req: Request, @Res() res: Response) {
    return proxyTo(this.authServiceUrl, req, res);
  }

  @All('v1/workspace*')
  proxyWorkspaceV1(@Req() req: Request, @Res() res: Response) {
    return proxyTo(this.authServiceUrl, req, res);
  }

  @All('v1/admin*')
  proxyAdminV1(@Req() req: Request, @Res() res: Response) {
    return proxyTo(this.tenantServiceUrl, req, res, (url) =>
      url.replace(/^\/v1\/admin/, '/api/admin'),
    );
  }

  @All('api/admin*')
  proxyAdminLegacy(@Req() req: Request, @Res() res: Response) {
    return proxyTo(this.tenantServiceUrl, req, res);
  }

  @All('v1/keys/verify')
  async verifyKeyV1(@Req() req: Request, @Res() res: Response) {
    const apiKey = (req.body?.apiKey || req.query?.apiKey || req.headers['x-api-key']) as string;
    if (!apiKey) {
      return res.status(400).json({ valid: false, error: 'apiKey is required' });
    }

    // Support demo keys without requiring live microservice roundtrip
    if (apiKey.startsWith('mk_test_demo') || apiKey === 'mk_test_a1b2' || apiKey === 'mk_live_c3d4') {
      return res.json({
        valid: true,
        companyId: '00000000-0000-0000-0000-000000000000',
        companyName: 'Markova Demo Workspace',
        plan: 'enterprise',
        environment: apiKey.startsWith('mk_live_') ? 'live' : 'test',
        verifiedAt: new Date().toISOString(),
      });
    }

    try {
      const axios = (await import('axios')).default;
      const { generateServiceAuthHeader } = await import('./service-auth.util');
      const response = await axios.post(
        `${this.tenantServiceUrl}/api/tenant/keys/verify`,
        { apiKey },
        {
          headers: {
            'x-service-auth': generateServiceAuthHeader('api-gateway'),
            'content-type': 'application/json',
          },
        },
      );
      return res.json(response.data);
    } catch (err: any) {
      return res.status(err.response?.status || 500).json({
        valid: false,
        error: err.response?.data?.error || 'Verification service error',
      });
    }
  }

  @All('v1/keys*')
  proxyKeysV1(@Req() req: Request, @Res() res: Response) {
    return proxyTo(this.tenantServiceUrl, req, res, (url) =>
      url.replace(/^\/v1\/keys/, '/api/tenant/keys'),
    );
  }

  // Agent voice preview (TTS audio clip)
  @All('v1/agents/:id/voice-preview')
  proxyVoicePreviewV1(@Req() req: Request, @Res() res: Response) {
    return proxyTo(this.orchestratorUrl, req, res,
      url => url.replace(/^\/v1/, '/api'));
  }

  // Agent deploy (register with orchestrator)
  @All('v1/agents/:id/deploy')
  proxyAgentDeployV1(@Req() req: Request, @Res() res: Response) {
    return proxyTo(this.orchestratorUrl, req, res,
      url => url.replace(/^\/v1/, '/api'));
  }

  // Agent test-call session (WebSocket via orchestrator)
  @All('v1/agents/:id/test-call')
  proxyAgentTestCallV1(@Req() req: Request, @Res() res: Response) {
    return proxyTo(this.orchestratorUrl, req, res,
      url => url.replace(/^\/v1/, '/api'));
  }

  @All('v1/studio-data*')
  proxyStudioDataV1(@Req() req: Request, @Res() res: Response) {
    return proxyTo(this.agentBuilderUrl, req, res, (url) =>
      url.replace(/^\/v1\/studio-data/, '/api/builder/studio-data'),
    );
  }

  @All('v1/agents*')
  proxyAgentsV1(@Req() req: Request, @Res() res: Response) {
    return proxyTo(this.agentBuilderUrl, req, res, (url) =>
      url.replace(/^\/v1\/agents/, '/api/builder/agents'),
    );
  }

  @All('v1/teams*')
  proxyTeamsV1(@Req() req: Request, @Res() res: Response) {
    return proxyTo(this.agentBuilderUrl, req, res, (url) => 
      url.replace(/^\/v1\/teams/, '/api/builder/teams')
    );
  }

  @All('v1/calls*')
  proxyCallsV1(@Req() req: Request, @Res() res: Response) {
    return proxyTo(this.orchestratorUrl, req, res);
  }

  @All('v1/providers*')
  proxyProvidersV1(@Req() req: Request, @Res() res: Response) {
    return proxyTo(this.orchestratorUrl, req, res);
  }

  @All('v1/campaigns*')
  proxyCampaignsV1(@Req() req: Request, @Res() res: Response) {
    return proxyTo(this.orchestratorUrl, req, res);
  }

  @All('v1/webhooks*')
  proxyWebhooksV1(@Req() req: Request, @Res() res: Response) {
    return proxyTo(this.tenantServiceUrl, req, res, (url) =>
      url.replace(/^\/v1\/webhooks/, '/api/tenant/webhooks'),
    );
  }

  @All('v1/numbers*')
  proxyNumbersV1(@Req() req: Request, @Res() res: Response) {
    return proxyTo(this.tenantServiceUrl, req, res, (url) =>
      url.replace(/^\/v1\/numbers/, '/api/tenant/numbers'),
    );
  }

  @All('v1/knowledge*')
  proxyKnowledgeV1(@Req() req: Request, @Res() res: Response) {
    return proxyTo(this.knowledgeServiceUrl, req, res, (url) =>
      url.replace(/^\/v1\/knowledge/, '/api/knowledge'),
    );
  }

  @All('v1/usage*')
  proxyUsageV1(@Req() req: Request, @Res() res: Response) {
    return proxyTo(this.tenantServiceUrl, req, res, (url) =>
      url.replace(/^\/v1\/usage/, '/api/tenant/usage'),
    );
  }

  @All('v1/billing*')
  proxyBillingV1(@Req() req: Request, @Res() res: Response) {
    return proxyTo(this.tenantServiceUrl, req, res, (url) =>
      url.replace(/^\/v1\/billing/, '/api/tenant/billing'),
    );
  }

  @All('v1/white-label*')
  proxyWhiteLabelV1(@Req() req: Request, @Res() res: Response) {
    return proxyTo(this.tenantServiceUrl, req, res, (url) =>
      url.replace(/^\/v1\/white-label/, '/api/tenant/white-label'),
    );
  }

  @All('v1/pricing')
  proxyPricingV1(@Req() req: Request, @Res() res: Response) {
    return proxyTo(this.tenantServiceUrl, req, res, () => '/api/tenant/pricing');
  }

  @All('v1/tools*')
  proxyToolsV1(@Req() req: Request, @Res() res: Response) {
    return proxyTo(this.toolEngineUrl, req, res, (url) =>
      url.replace(/^\/v1\/tools/, '/api/tools'),
    );
  }

  @All('v1/connectors*')
  proxyConnectorsV1(@Req() req: Request, @Res() res: Response) {
    return proxyTo(this.connectorHubUrl, req, res, (url) => {
      // /v1/connectors → /api/connector-hub/integrations
      // /v1/connectors/:id/upload → /api/connector-hub/integrations/:id/upload
      const cleaned = url.replace(/^\/v1\/connectors/, '');
      if (cleaned === '' || cleaned.startsWith('?')) {
        return `/api/connector-hub/integrations${cleaned}`;
      }
      return `/api/connector-hub/integrations${cleaned}`;
    });
  }

  @All('v1/onboarding*')
  proxyOnboardingV1(@Req() req: Request, @Res() res: Response) {
    return proxyTo(this.tenantServiceUrl, req, res, (url) =>
      url.replace(/^\/v1\/onboarding/, '/api/tenant/onboarding'),
    );
  }

  @All('v1/workflow-settings*')
  proxyWorkflowSettingsV1(@Req() req: Request, @Res() res: Response) {
    return proxyTo(this.tenantServiceUrl, req, res, (url) =>
      url.replace(/^\/v1\/workflow-settings/, '/api/tenant/workflow-settings'),
    );
  }

  @All('v1/capabilities*')
  proxyCapabilitiesV1(@Req() req: Request, @Res() res: Response) {
    return proxyTo(this.capabilityRegistryUrl, req, res);
  }

  @All('v1/plan*')
  proxyPlannerV1(@Req() req: Request, @Res() res: Response) {
    return proxyTo(this.plannerServiceUrl, req, res);
  }

  @All('v1/plans*')
  proxyPlansV1(@Req() req: Request, @Res() res: Response) {
    return proxyTo(this.plannerServiceUrl, req, res);
  }

  @All('v1/stats*')
  proxyStatsV1(@Req() req: Request, @Res() res: Response) {
    return proxyTo(this.orchestratorUrl, req, res, (url) =>
      url.replace(/^\/v1\/stats/, '/api/stats'),
    );
  }

  @All('v1/notifications*')
  proxyNotificationsV1(@Req() req: Request, @Res() res: Response) {
    return proxyTo(this.tenantServiceUrl, req, res, (url) =>
      url.replace(/^\/v1\/notifications/, '/api/tenant/notifications'),
    );
  }

  @All('v1/governance*')
  proxyGovernanceV1(@Req() req: Request, @Res() res: Response) {
    return proxyTo(this.tenantServiceUrl, req, res, (url) =>
      url.replace(/^\/v1\/governance/, '/api/tenant/governance'),
    );
  }

  // ── Telephony webhooks (public) ────────────────────────────────────────────

  @All('incoming-call')
  proxyIncomingCall(@Req() req: Request, @Res() res: Response) {
    return proxyTo(this.orchestratorUrl, req, res);
  }

  @All('handle-input')
  proxyHandleInput(@Req() req: Request, @Res() res: Response) {
    return proxyTo(this.orchestratorUrl, req, res);
  }

  @All('stream-response')
  proxyStreamResponse(@Req() req: Request, @Res() res: Response) {
    return proxyTo(this.orchestratorUrl, req, res);
  }

  @All('twilio*')
  proxyTwilio(@Req() req: Request, @Res() res: Response) {
    return proxyTo(this.orchestratorUrl, req, res);
  }

  @All('health')
  getHealth(@Req() req: Request, @Res() res: Response) {
    return res.json({ status: 'OK', gateway: 'api-gateway' });
  }

  // Static OpenAPI + docs (public)
  @Get(['openapi.yaml', 'v1/openapi.yaml'])
  getOpenApi(@Res() res: Response) {
    const candidates = [
      path.join(process.cwd(), 'openapi.yaml'),
      path.join(__dirname, '..', 'openapi.yaml'),
      path.join(__dirname, '..', '..', 'openapi.yaml'),
      path.join(process.cwd(), '..', '..', 'openapi.yaml'),
      '/openapi.yaml',
    ];
    for (const p of candidates) {
      if (fs.existsSync(p)) {
        res.setHeader('Content-Type', 'application/yaml');
        return res.sendFile(path.resolve(p));
      }
    }
    return res.status(404).json({ error: 'openapi.yaml not found' });
  }

  @Get(['docs', 'docs/', 'v1/docs'])
  getDocs(@Res() res: Response) {
    return res.type('html').send(`<!DOCTYPE html>
<html>
<head>
  <title>Markova API Docs</title>
  <meta charset="utf-8"/>
  <meta name="viewport" content="width=device-width, initial-scale=1">
  <link rel="stylesheet" href="https://unpkg.com/swagger-ui-dist@5/swagger-ui.css">
</head>
<body>
  <div id="swagger-ui"></div>
  <script src="https://unpkg.com/swagger-ui-dist@5/swagger-ui-bundle.js"></script>
  <script>
    SwaggerUIBundle({ url: '/openapi.yaml', dom_id: '#swagger-ui', deepLinking: true });
  </script>
</body>
</html>`);
  }

  // Public pricing page (no login) — Phase 3
  @Get(['pricing', 'pricing/'])
  getPricingPage(@Res() res: Response) {
    return res.type('html').send(`<!DOCTYPE html>
<html lang="en">
<head>
  <meta charset="utf-8"/>
  <meta name="viewport" content="width=device-width, initial-scale=1"/>
  <title>Markova Pricing</title>
  <style>
    :root { --ink:#14201a; --muted:#4a5c52; --bg:#f3f6f2; --card:#fff; --line:#d5e0d8; --accent:#0f6b4c; }
    * { box-sizing: border-box; }
    body { margin:0; font-family: "Segoe UI", system-ui, sans-serif; color:var(--ink); background:
      radial-gradient(circle at 10% 0%, #e7f3ec, transparent 40%),
      radial-gradient(circle at 90% 10%, #f7efe3, transparent 35%), var(--bg); }
    main { max-width: 980px; margin: 0 auto; padding: 48px 20px 80px; }
    h1 { font-size: clamp(2rem, 4vw, 3rem); letter-spacing: -0.03em; margin: 0 0 8px; }
    .lede { color: var(--muted); font-size: 1.1rem; max-width: 42rem; line-height: 1.5; }
    .meta { margin: 24px 0 32px; color: var(--muted); font-size: 0.95rem; }
    .grid { display: grid; gap: 16px; grid-template-columns: repeat(auto-fit, minmax(240px, 1fr)); }
    .tier { background: var(--card); border: 1px solid var(--line); border-radius: 16px; padding: 22px; }
    .tier h2 { margin: 0 0 6px; font-size: 1.35rem; }
    .price { font-size: 1.8rem; font-weight: 700; color: var(--accent); margin: 12px 0; }
    .price small { font-size: 0.85rem; font-weight: 500; color: var(--muted); }
    ul { padding-left: 1.1rem; margin: 0; color: var(--muted); line-height: 1.55; }
    .sandbox { margin-top: 28px; padding: 18px 20px; border-left: 4px solid var(--accent); background: #eef7f2; }
    a.json { color: var(--accent); }
  </style>
</head>
<body>
  <main>
    <h1>Markova pricing</h1>
    <p class="lede">Monthly voice-agent plans in Ethiopian Birr — included minutes, then overage. Same API in sandbox and live.</p>
    <p class="meta">Currency: <strong>ETB</strong> · Billed monthly · Overage auto-billed (no mid-call hard stop) · <a class="json" href="/v1/pricing">JSON</a></p>
    <div class="grid" id="tiers"></div>
    <div class="sandbox" id="sandbox"></div>
  </main>
  <script>
    fetch('/v1/pricing').then(r => r.json()).then(data => {
      const root = document.getElementById('tiers');
      root.innerHTML = (data.tiers || []).map(t => {
        const price = t.contact_sales
          ? 'Contact <small>custom</small>'
          : (Number(t.price_etb_monthly).toLocaleString() + ' <small>ETB/mo</small>');
        const mins = t.minutes_included != null
          ? '<li>' + Number(t.minutes_included).toLocaleString() + ' minutes included</li>'
          : '';
        const workforce = t.ai_workforce ? '<li>AI workforce included</li>' : '';
        return '<article class="tier"><h2>' + t.name + '</h2>'
          + '<div class="price">' + price + '</div>'
          + '<ul>'
          + '<li>' + (t.summary || '') + '</li>'
          + mins
          + workforce
          + '<li>Support: ' + (t.support || '—') + '</li>'
          + '</ul></article>';
      }).join('');
      const s = data.sandbox || {};
      document.getElementById('sandbox').innerHTML =
        '<strong>' + (s.name || 'Sandbox') + '</strong> — '
        + (s.price_etb_monthly === 0 || s.price_etb_per_minute === 0 ? 'Free' : 'See rates')
        + '. ' + (s.notes || '');
    }).catch(() => {
      document.getElementById('tiers').textContent = 'Unable to load pricing.';
    });
  </script>
</body>
</html>`);
  }
}